+++ packages/terra-core-docs/src/terra-dev-site/test/status-view/StatusViewOverflow.test.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import classNames from 'classnames/bind';
+import StatusView from 'terra-status-view';
+import styles from './StatusViewTestCommon.module.scss';
+
+const cx = classNames.bind(styles);
+
+export default () => (
+  <div className={cx('status-view-wrapper-3')}>
+    <StatusView id="statusView" variant="no-data" className={cx('status-view-wrapper-3')} />
+  </div>
+);