+++ packages/terra-section-header/src/terra-dev-site/doc/example/DefaultSectionHeader.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import SectionHeaderExampleTemplate from 'terra-section-header/lib/terra-dev-site/doc/example/SectionHeaderExampleTemplate';
-
-const DefaultSectionHeader = () => <SectionHeaderExampleTemplate title="Default" exampleProps={{ title: 'Default Section Header' }} />;
-
-export default DefaultSectionHeader;