+++ packages/terra-status-view/src/terra-dev-site/test/status-view/StatusViewNotAuthorized.test.jsx
@@ -1,6 +1,0 @@
-import React from 'react';
-import StatusView from '../../../StatusView';
-
-export default () => (
-  <StatusView id="statusView" variant="not-authorized" />
-);